requirements.txt
Django==2.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
python-dotenv==0.7.1
